WIP Project Carousel
I changed the About Me section to include a profile picture. I'm revamping the projects section (wip). Organized components into their respective folder
