+++ package.json
@@ -20,7 +20,8 @@
     "portfolio": "file:",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.5",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "swiper": "^11.1.15"
   },
   "devDependencies": {
     "@eslint/js": "^9.11.1",
