+++ src/App.css
@@ -74,20 +74,17 @@ a {
   margin: 0;
 }
 
+.soft-blue{
+  color: rgb(157 249 217);
+}
+
 a {
   cursor: pointer;
 }
 
-.bg-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.content {
-  position: relative;
-  z-index: 1;
-  max-width: 50%;
+.content{
+  max-width: 50vw;
+  
 }
 
 .fade-in {
@@ -99,6 +96,16 @@ a {
 .fade-in.is-visible {
   opacity: 1;
   transform: translateY(0);
+}
+
+.swiper-button-prev,
+.swiper-button-next {
+  height: 40px;
+  background-color: rgba(235, 15, 15, 0.5); /* Optional: Add background for better visibility */
+  border-radius: 50%; /* Make it circular */
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
 @media screen and (max-width: 760px) {
