+++ src/components/Skills/SkillBadge.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
-import styles from "../assets/styles/Skills.module.css";
+import styles from "../../assets/styles/Skills.module.css";
 
 function SkillBadge({
   bgColor = "rgba(255,255,255,0.3)",
